refactor strategy
